Add test cases for intermediate variable (#33)
I was curious about how this plugin treated code that assigns propTypes
to an intermediate variable. I'm not entirely sure if this is the
desired affect, but I figured I would add a couple of test cases that
demonstrate this.

I wanted to test exported and non-exported variables separately in case
there was a difference, but it turned out to not have much difference.

One possible avenue for improvement would be for non-exported, unused,
intermediate variables to be removed--but perhaps that is the domain of
a separate plugin.