Remove newly unreferenced VariableDeclarators
I've noticed that a common pattern used to assign propTypes--first
define a variable for the propTypes, and then assign it to the component
by reference--ends up deoptimizing this plugin, which leads to larger
bundles in production than desired. The original hypothesis that
minification would clean up these dead references works well for the
simplest of propTypes, but unfortunately it doesn't hold up as
soon as you have a function call in the propTypes variable (such as is
the case for propTypes like `arrayOf`, `objectOf`, `shape`, etc.)
because the minfier does not know that the function call will not have
side-effects:

```js
const propTypes = {
  foo: PropTypes.arrayOf(PropTypes.number),
};
```

I believe this should be solved by this plugin because we can make the
assumption that function calls inside of propTypes have no side-effects.

There seems to be a weird issue where the transform adds a semicolon
when using in wrapped mode. For now, I decided to simply add the
additional semicolons in the test fixtures, since they will be minified
out anyway.

I don't have much experience working with Babel plugins yet, so it is
likely that I'm doing things in a weird or sub-optimal way here.